Create README.md
I've been studying programming language for 3 months in the front end development area. I have a good knowledge of HTML and CSS. In javascript, I'm still at a basic level, as I only do a few weeks of study in the language. This profile is periodically updated. My goal is to become a full stack web developer.